feat: Add is_inside_working_tree method to RepoDir
This commit adds a new method `is_inside_working_tree` to the `RepoDir` class.

This method checks if the Git repository is a non-bare repository (i.e., has a working tree) by utilizing `git rev-parse --is-bare-repository`.

The `__slots__` attribute and the `main` function have been updated to reflect this change.